chore: refactor add-edit-connetion-button
